Break pokedex stylings into smaller components
